Implemented graceful denial of voting for guest users.

+++ app/controllers/application_controller.rb
@@ -8,62 +8,65 @@ class ApplicationController < ActionController::Base
   end
 
   def vote
+    if can? :vote, current_user
+      direction = params[:vote]
 
-    direction = params[:vote]
-
-    if params[:votable] == "comment"
-      votable = Comment.find(params[:id])
-    else
-      votable = Proposal.find(params[:id])
-    end
-
-    author = votable.user
-
-    voting = Voting.find_or_initialize_by({
-      user: current_user,
-      votable: votable
-    })
-
-    # voting is initialized
-    if voting.value.nil?
-
-      voting.value = Voting.values[direction.to_sym]
-      voting.save
-
-      direction == "up" ? votable.up += 1 : votable.down += 1
-      author.reputation += direction == "up" ? 1 : -1
-
-    # voting is already created and user REMOVED her/his vote
-    elsif voting.value == Voting.values[direction.to_sym]
-
-      voting.destroy
-
-      direction == "up" ? votable.up -= 1 : votable.down -= 1
-      author.reputation += direction == "up" ? -1 : 1
-
-    # voting is already created and user CHANGED her/his vote
-    elsif voting.value != Voting.values[direction.to_sym]
-
-      voting.value = Voting.values[direction.to_sym]
-      voting.save
-
-      if direction == "up"
-        votable.up += 1;
-        votable.down -= 1
+      if params[:votable] == "comment"
+        votable = Comment.find(params[:id])
       else
-        votable.up -= 1;
-        votable.down += 1
+        votable = Proposal.find(params[:id])
       end
 
-      author.reputation += direction == "up" ? 2 : -2
+      author = votable.user
 
+      voting = Voting.find_or_initialize_by({
+        user: current_user,
+        votable: votable
+      })
+
+      # voting is initialized
+      if voting.value.nil?
+
+        voting.value = Voting.values[direction.to_sym]
+        voting.save
+
+        direction == "up" ? votable.up += 1 : votable.down += 1
+        author.reputation += direction == "up" ? 1 : -1
+
+      # voting is already created and user REMOVED her/his vote
+      elsif voting.value == Voting.values[direction.to_sym]
+
+        voting.destroy
+
+        direction == "up" ? votable.up -= 1 : votable.down -= 1
+        author.reputation += direction == "up" ? -1 : 1
+
+      # voting is already created and user CHANGED her/his vote
+      elsif voting.value != Voting.values[direction.to_sym]
+
+        voting.value = Voting.values[direction.to_sym]
+        voting.save
+
+        if direction == "up"
+          votable.up += 1;
+          votable.down -= 1
+        else
+          votable.up -= 1;
+          votable.down += 1
+        end
+
+        author.reputation += direction == "up" ? 2 : -2
+
+      end
+
+      votable.hotness = votable.up - votable.down
+      votable.save
+      author.save
+
+      render json: { hotness: votable.hotness }
+    else
+      render json: { error: "Unauthorized" }, status: :unprocessable_entity
     end
-
-    votable.hotness = votable.up - votable.down
-    votable.save
-    author.save
-
-    render json: { hotness: votable.hotness }
 
   end
